+++ core/wrappered_graph.hpp
@@ -104,7 +104,7 @@ namespace mcmc_utilities
       auto p=std::dynamic_pointer_cast<stochastic_node<T,std_vector> >(pn);
       if(p)
 	{
-	  for(int i=0;i<v.size();++i)
+	  for(size_t i=0;i<v.size();++i)
 	    {
 	      p->set_value(i,v[i]);
 	    }
@@ -127,7 +127,7 @@ namespace mcmc_utilities
       auto p=std::dynamic_pointer_cast<stochastic_node<T,std_vector> >(pn);
       if(p)
 	{
-	  for(int i=0;i<v.size();++i)
+	  for(size_t i=0;i<v.size();++i)
 	    {
 	      p->set_value(i,v[i]);
 	      p->set_observed(i,true);
